+++ src/components/inputs/ElTextInput.vue
@@ -1,7 +1,7 @@
 <template>
 	<b-form-input
-	:value="value"
-	@input="update($event.target.value)" />
+		:value="value"
+		@input="update($event.target.value)" />
 </template>
 
 <script>
