+++ src/index.js
@@ -4,12 +4,18 @@ import BootstrapVue from 'bootstrap-vue';
 import scrollTo from 'vue-scrollto';
 
 import ElButton from './components/ElButton.vue';
-import ElCodeInput from './components/ElCodeInput.vue';
 import ElError from './components/ElError.vue';
+import ElForm from './components/ElForm.vue';
 import ElHeader from './components/ElHeader.vue';
 import ElLayout from './components/ElLayout.vue';
 import ElMarkdown from './components/ElMarkdown.vue';
 import ElTaskbar from './components/ElTaskbar.vue';
+
+import ElCodeInput from './components/inputs/ElCodeInput';
+import ElForm from './components/inputs/ElForm';
+import ElSelectInput from './components/inputs/ElSelectInput';
+import ElTextAreaInput from './components/inputs/ElTextAreaInput';
+import ElTextInput from './components/inputs/ElTextInput';
 
 Vue.use(BootstrapVue);
 Vue.use(scrollTo, {
@@ -19,12 +25,17 @@ Vue.use(scrollTo, {
 
 const components = {
 	ElButton,
-	ElCodeInput,
 	ElError,
+	ElForm,
 	ElHeader,
 	ElLayout,
 	ElMarkdown,
-	ElTaskbar
+	ElTaskbar,
+	ElCodeInput,
+	ElForm,
+	ElSelectInput,
+	ElTextAreaInput,
+	ElTextInput
 };
 
 export default function install($Vue) {
